+++ test/bedrock/10-api.test.js
@@ -70,8 +70,7 @@ const testWorkflow = {
   }),
   clientSecret: 'shhh',
   oidc: {
-    redirectUri: 'https://example.com',
-    scopes: [{name: 'openid'}]
+    redirectUri: 'https://example.com'
   }
 };