+++ test/bedrock/200-18013-7-native.test.js
@@ -46,8 +46,7 @@ const mdocTestRP = {
   }],
   clientSecret: 'shhh',
   oidc: {
-    redirectUri: 'https://example.com',
-    scopes: [{name: 'openid'}]
+    redirectUri: 'https://example.com'
   }
 };
 
@@ -69,8 +68,7 @@ const mixedFormatTestRP = {
   ],
   clientSecret: 'shhh',
   oidc: {
-    redirectUri: 'https://example.com',
-    scopes: [{name: 'openid'}]
+    redirectUri: 'https://example.com'
   }
 };