+++ test/bedrock/220-oid4vp-x509.test.js
@@ -44,8 +44,7 @@ const testWorkflow = {
   ],
   clientSecret: 'shhh',
   oidc: {
-    redirectUri: 'https://example.com',
-    scopes: [{name: 'openid'}]
+    redirectUri: 'https://example.com'
   }
 };